requirements.txt
aioodbc
aiosmtplib
annotated-types
anyio
asyncpg
attrs
bcrypt
bidict
blinker
CacheControl
cachetools
certifi
cffi
charset-normalizer
click
colorama
cryptography
ecdsa
fastapi
filelock
firebase
firebase-admin
gitdb
GitPython
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-core
google-cloud-firestore
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
httpcore
httplib2
httpx
idna
Jinja2
jsonschema
jsonschema-specifications
MarkupSafe
msgpack
narwhals
packaging
pillow
proto-plus
protobuf
psycopg2-binary
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
Pygments
PyJWT
pyodbc
pyparsing
python-dateutil
python-dotenv
python-engineio
python-jose
python-multipart
python-socketio
pytz
referencing
requests
rich
rpds-py
rsa
setuptools
simple-websocket
six
smmap
sniffio
SQLAlchemy
starlette
tenacity
termcolor
toml
tornado
typing-inspection
typing_extensions
tzdata
uritemplate
urllib3
uuid
uvicorn
watchdog
Werkzeug
wheel
wrapt
wsproto
